make fixture "config dir" global
