+++ tests/pipeline/test_pipeline_ml.py
@@ -1,9 +1,4 @@
-from copy import deepcopy
-from pathlib import Path
-from typing import Dict
-
 import pytest
-import yaml
 from kedro import __version__ as KEDRO_VERSION
 from kedro.extras.datasets.pandas import CSVDataSet
 from kedro.framework.context import KedroContext
@@ -16,57 +11,12 @@ from kedro_mlflow.pipeline import KedroMlflowPipelineMLInputsError, PipelineML
 def _write_yaml(filepath: Path, config: Dict):
     filepath.parent.mkdir(parents=True, exist_ok=True)
     yaml_str = yaml.dump(config)
-    filepath.write_text(yaml_str)
-
-
-def _get_local_logging_config():
-    return {
-        "version": 1,
-        "formatters": {
-            "simple": {"format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s"}
-        },
-        "root": {"level": "INFO", "handlers": ["console"]},
-        "loggers": {
-            "kedro": {"level": "INFO", "handlers": ["console"], "propagate": False}
-        },
-        "handlers": {
-            "console": {
-                "class": "logging.StreamHandler",
-                "level": "INFO",
-                "formatter": "simple",
-                "stream": "ext://sys.stdout",
-            }
-        },
-        "info_file_handler": {
-            "class": "logging.handlers.RotatingFileHandler",
-            "level": "INFO",
-            "formatter": "simple",
-            "filename": "logs/info.log",
-        },
-    }
-
-
-@pytest.fixture
-def config_dir(tmp_path):
-    """This emulates the root of a kedro project.
-    This function must be called before any instantiation of DummyContext
-    """
-    for env in ["base", "local"]:
-        catalog = tmp_path / "conf" / env / "catalog.yml"
-        credentials = tmp_path / "conf" / env / "credentials.yml"
-        logging = tmp_path / "conf" / env / "logging.yml"
-        parameters = tmp_path / "conf" / env / "parameters.yml"
-        _write_yaml(catalog, dict())
-        _write_yaml(parameters, dict())
-        _write_yaml(credentials, dict())
-        _write_yaml(logging, _get_local_logging_config())
 
 
 def preprocess_fun(data):
     return data
 
 
-def train_fun(data):
     return 2
 
 
@@ -119,6 +69,17 @@ def dummy_context(tmp_path, config_dir, mocker):
     return DummyContext(tmp_path.as_posix())
 
 
+@pytest.fixture
+def dummy_catalog():
+    return DataCatalog(
+        {
+            "raw_data": MemoryDataSet(),
+            "data": MemoryDataSet(),
+            "model": CSVDataSet("fake/path/to/file.csv"),
+        }
+    )
+
+
 @pytest.mark.parametrize(
     "tags,from_nodes,to_nodes,node_names,from_inputs",
     [
@@ -147,9 +108,8 @@ def test_filtering_pipeline_ml(
      We loop dynamically on the arguments of the function in case of kedro
      modify the filters.
     """
-    # prevent checks when consructing PipelineML because artifacts are not persisted
-    # mocker.patch.object(PipelineML, "_check_degrees_of_freedom", new=lambda self: True)
-    # mocker.patch.object(PipelineML, "_check_model_input_name", new=lambda self, x: True)
+
+    # dummy_context, pipeline_with_tag, pipeline_ml_with_tag are fixture in conftest
 
     # remember : the arguments are iterable, so do not pass string directly (e.g ["training"] rather than training)
     filtered_pipeline = dummy_context._filter_pipeline(
@@ -223,15 +183,9 @@ def test_filtering_generate_invalid_pipeline_ml(
 #     pass
 
 # filtering that remove the degree of freedom constraints should fail
-def test_catalog_extraction(pipeline_ml_with_tag):
-    catalog = DataCatalog(
-        {
-            "raw_data": MemoryDataSet(),
-            "data": MemoryDataSet(),
-            "model": CSVDataSet("fake/path/to/file.csv"),
-        }
-    )
-    filtered_catalog = pipeline_ml_with_tag.extract_pipeline_catalog(catalog)
+def test_catalog_extraction(pipeline_ml_with_tag, dummy_catalog):
+
+    filtered_catalog = pipeline_ml_with_tag.extract_pipeline_catalog(dummy_catalog)
     assert set(filtered_catalog.list()) == {"model", "data"}
